MyFunction/__init__.py
import logging
from datetime import datetime

import azure.functions as func

from .topper_student import Student

Topper=Student("Tukaram", datetime(1999,2,15), "MECH", 20)

def main(req: func.HttpRequest) -> func.HttpResponse:
    logging.info('Python HTTP trigger function processed a request.')

    name = req.params.get('name')
    print("Hi all Good afternoon")
    if not name:
        try:
            req_body = req.get_json()
        except ValueError:
            pass
        else:
            name = req_body.get('name')

    if name:
        return func.HttpResponse(f"Hello, {name}. This HTTP triggered function executed successfully.So brillient student age  is {Topper.get_age()}")
    else:
        return func.HttpResponse(
             "This HTTP triggered function executed successfully. Pass a name in the query string or in the request body for a personalized response.",
             status_code=200
        )
